jekyll-sitemap uses &url to map sitemap.xml
We already have sitemap.txt which is just fine, but:

https://github.com/jekyll/jekyll-sitemap

```
url: "http://example.com" # the base hostname & protocol for your site
plugins:
  - jekyll-sitemap
```

```yml
url: "http://example.com" # the base hostname & protocol for your site
plugins:
  - jekyll-sitemap
  ```

  If all gem plugins have the same `priority`, they will be executed in
  the
  order they are required, generally. Thus, if you have other plugins
  which
  generate content and store that content in `site.pages`, `site.posts`,
  or
  `site.collections`, be sure to require `jekyll-sitemap` either *after*
  those other gems if you *want* the sitemap to include the generated
  content, or *before* those other gems if you *don't want* the sitemap
  to
  include the generated content from the gems. (Programming is *hard*.)

  Because the sitemap is added to `site.pages`, you may have to modify
  any
  templates that iterate through all pages (for example, to build a menu
  of
  all of the site's content).

 ## Note on Use with Github Pages Gem
The Github Pages gem ignores all plugins included in the Gemfile. If you
only include `jekyll-sitemap` in the Gemfile without also including it
in the `_config.yml` *the plugin will not work*. This can be confusing
because the official Jekyll docs state that plugins can be included in
either the Gemfile or `_config.yml`.

When building a site that uses the Github Pages gem, follow the
instructions above and ensure that `jekyll-sitemap` is listed in the
`gems` array in `_config.yml`.

 ## `<lastmod>` tag
The `<lastmod>` tag in the `sitemap.xml` will reflect by priority:

1.   The modified date of the file as reported by the filesystem if you
have `jekyll-last-modified-at` plugin installed (not compatible with
Github Pages auto building)
2.   A personalised date if you add the variable `last_modified_at:`
with a date in the Front Matter
3.   The creation date of your post (corresponding to the `post.date`
variable)

 ## Exclusions

If you would like to exclude specific pages/posts from the sitemap set
the
sitemap flag to `false` in the front matter for the page/post.

```yml
sitemap: false
```